% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_genelist_terms.R
\name{bar_genelist_terms}
\alias{bar_genelist_terms}
\title{Generate horizontal bar plot   of terms  count across gene list
This function generates a horizontal bar plot of term count across the gene
list for top 30 most significant terms based on combined score.
This bar plot visualizes how many terms (pathways) are distributed
in the gene lists.}
\usage{
bar_genelist_terms(enrichr_df, minimum_combined_score = 5)
}
\arguments{
\item{enrichr_df}{A data frame generated  by implementing the function
enrichr_df}

\item{minimum_combined_score}{A integer value below which the combined score
is set to be insignificant. Default is 5.}
}
\value{
A bar plot.
}
\description{
Generate horizontal bar plot   of terms  count across gene list
This function generates a horizontal bar plot of term count across the gene
list for top 30 most significant terms based on combined score.
This bar plot visualizes how many terms (pathways) are distributed
in the gene lists.
}
