% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_plot_count.R
\name{bar_plot_count}
\alias{bar_plot_count}
\title{Bar plot of overall term count
Generates bar plot of term count across the gene lists.
This bar plot visualizes how many terms (pathways) are distributed
in the gene lists.}
\usage{
bar_plot_count(enrichr_df, minimum_combined_score = 5)
}
\arguments{
\item{enrichr_df}{A data frame generated  by implementing the function
enrichr_df.}

\item{minimum_combined_score}{A integer value below which the combined score
is set to be insignificant.Default is 5.}
}
\value{
A bar plot.
}
\description{
Bar plot of overall term count
Generates bar plot of term count across the gene lists.
This bar plot visualizes how many terms (pathways) are distributed
in the gene lists.
}
