% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_plot_genelist_library.R
\name{bar_plot_genelist_library}
\alias{bar_plot_genelist_library}
\title{Composite Bar plot of overall term count
Generates a mixed bar plot of term count across the gene list and libraries.
This bar plot visualizes how many terms (pathways) are distributed.
in the gene lists.}
\usage{
bar_plot_genelist_library(enrichr_df, minimum_combined_score = 5)
}
\arguments{
\item{enrichr_df}{A data frame generated  by implementing the function enrichr_df.}

\item{minimum_combined_score}{A integer value below which the combined score is set to be insignificant. Default is 5.}
}
\value{
A bar plot.
}
\description{
Composite Bar plot of overall term count
Generates a mixed bar plot of term count across the gene list and libraries.
This bar plot visualizes how many terms (pathways) are distributed.
in the gene lists.
}
