% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_enrichR_lib.R
\name{db_EnrichR_lib}
\alias{db_EnrichR_lib}
\title{Extract list of currently available db in EnrichR
This function extracts the list  of database currently available in EnrichR.
It does not have parameter. It implements listEnrichrDbs functionality
hosted in enrichr package. NOTE: enrichr package must  be installed  to
implement this function.}
\usage{
db_EnrichR_lib()
}
\value{
An .xlsx file in data dir. Creates data dir if does not exist.
}
\description{
Extract list of currently available db in EnrichR
This function extracts the list  of database currently available in EnrichR.
It does not have parameter. It implements listEnrichrDbs functionality
hosted in enrichr package. NOTE: enrichr package must  be installed  to
implement this function.
}
