% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_venn.R.R
\name{draw_venn_diagram_terms}
\alias{draw_venn_diagram_terms}
\title{This function creates the venn diagram.
draw.venn(}
\usage{
draw_venn_diagram_terms(enrichr_df, term_list)
}
\description{
This function creates the venn diagram.
draw.venn(
}
