% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichr_df.R
\name{enrichr_df}
\alias{enrichr_df}
\title{Generate data frame from EnrichR output
This function generates a data frame with 9 features.  Make sure
implement_enrichr.R is implemented and outputs are in enrichr_results
directory.}
\usage{
enrichr_df()
}
\value{
A data frame
}
\description{
Generate data frame from EnrichR output
This function generates a data frame with 9 features.  Make sure
implement_enrichr.R is implemented and outputs are in enrichr_results
directory.
}
