% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_map.R
\name{enrichr_heat_map}
\alias{enrichr_heat_map}
\title{Generate heat map
It generates the heat map of the most frequent gene found in the database.}
\usage{
enrichr_heat_map(expanded_enrichr_df, minimum_combined_score = 5)
}
\arguments{
\item{expanded_enrichr_df}{A data frame generated  by exploding  the overlapping genes of query gene
list and database gene list. This data frame is created by implementing
function expanded_enrichr_df.}

\item{minimum_combined_score}{A integer value below which the combined score is set to be insignificant.
Default is 5.}
}
\value{
A heat map
}
\description{
Generate heat map
It generates the heat map of the most frequent gene found in the database.
}
