% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_table_enrichr_ubiquitous_terms.R
\name{enrichr_ubiquitous_terms}
\alias{enrichr_ubiquitous_terms}
\title{Generate table of the most common terms or pathways
It generates the list  of terms (could be a single term) that has the  highest
count across gene lists and libraries.}
\usage{
enrichr_ubiquitous_terms(enrichr_df, minimum_combined_score = 5)
}
\arguments{
\item{enrichr_df}{A dataframe generated  by implementing the function
enrichr_df}

\item{minimum_combined_score}{A integer value below which the combined score
is set to be insignificant. Default is 5.}
}
\value{
An excel file in data dir.
}
\description{
Generate table of the most common terms or pathways
It generates the list  of terms (could be a single term) that has the  highest
count across gene lists and libraries.
}
