% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_table_unique_terms.R
\name{enrichr_unique_terms}
\alias{enrichr_unique_terms}
\title{Generate of table of the unique terms or pathways
It generates the list  of terms that  occur only one time across gene list
and libraries.}
\usage{
enrichr_unique_terms(enrichr_df, minimum_combined_score = 5)
}
\arguments{
\item{enrichr_df}{A data frame generated  by implementing the function enrichr_df}

\item{minimum_combined_score}{A integer value below which the combined score is set to be insignificant}
}
\value{
A table in data directory.
}
\description{
Generate of table of the unique terms or pathways
It generates the list  of terms that  occur only one time across gene list
and libraries.
}
