% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_jaccard_similarity.R
\name{estimate_jaccard_similarity}
\alias{estimate_jaccard_similarity}
\title{Estimate Jaccard's similarity
This function estimates Jaccard's similarity between users' gene set and
database gene set.}
\usage{
estimate_jaccard_similarity(enrichr_data_frame)
}
\arguments{
\item{enrichr_data_frame}{The data frame generated by function  enrichr_df.}
}
\value{
The input  data frame is added  with a new column with Jaccard
similarity estimate.
}
\description{
Estimate Jaccard's similarity
This function estimates Jaccard's similarity between users' gene set and
database gene set.
}
