% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expanded_enrichr_df.R
\name{expanded_enrichr_df}
\alias{expanded_enrichr_df}
\title{Generate dataframe from EnrichR output
This function generates a dataframe with 9 features. It requires the
.xlsx output of implement_enrichr.R implementation.}
\usage{
expanded_enrichr_df()
}
\value{
A data frame
}
\description{
Generate dataframe from EnrichR output
This function generates a dataframe with 9 features. It requires the
.xlsx output of implement_enrichr.R implementation.
}
