% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Venn_diagram.R
\name{gene_list_venn_diagram}
\alias{gene_list_venn_diagram}
\title{Venn diagram
This function draws venn diagram}
\usage{
gene_list_venn_diagram(gene_list_file, gene_set_names)
}
\description{
Venn diagram
This function draws venn diagram
}
