% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implement_enrichr.R
\name{implement_enrichr}
\alias{implement_enrichr}
\title{Implement EnrichR
This function implements listEnrichrDBS and enrichr functions of EnrichR. It
requires the list of database to be explored and the csv file with gene_list.}
\usage{
implement_enrichr(dbs, gene_list)
}
\arguments{
\item{dbs}{A list of database}

\item{gene_list}{a csv file of the gene set}
}
\value{
An directory named enrichR results in data directory.
}
\description{
Implement EnrichR
This function implements listEnrichrDBS and enrichr functions of EnrichR. It
requires the list of database to be explored and the csv file with gene_list.
}
