% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_plot.R
\name{tile_plot}
\alias{tile_plot}
\title{Generate tile map
It generates the tile of the most frequent gene found in the database. This
plot is helpful to assess which library has more  coverage of genes for each
gene list.}
\usage{
tile_plot(expanded_enrichr_df, minimum_combined_score = 5)
}
\arguments{
\item{expanded_enrichr_df}{A data frame generated  by exploding  the
overlapping genes of query gene list and database gene list. This data frame
is created by implementing function expanded_enrichr_df.}

\item{minimum_combined_score}{An integer value below which the combined score
is set to be insignificant. Default is 5.}
}
\value{
A tile plot.
}
\description{
Generate tile map
It generates the tile of the most frequent gene found in the database. This
plot is helpful to assess which library has more  coverage of genes for each
gene list.
}
