% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/violin_plot_genelist.R
\name{violin_plot_genelist}
\alias{violin_plot_genelist}
\title{Violin plot of gene list and library
Generates violin  plot for each gene list for each library. Maximum number
of gene list and library recommended for this visualization is 10 each.}
\usage{
violin_plot_genelist(enrichr_df, minimum_combined_score = 5)
}
\arguments{
\item{enrichr_df}{A data frame generated  by implementing the function enrichr_df}

\item{minimum_combined_score}{A integer value below which the combined score
is set to be insignificant}
}
\value{
A violin plot
}
\description{
Violin plot of gene list and library
Generates violin  plot for each gene list for each library. Maximum number
of gene list and library recommended for this visualization is 10 each.
}
